+++ core/src/main/java/org/eqasim/core/simulation/mode_choice/EqasimModeChoiceModule.java
@@ -27,8 +27,8 @@ import org.eqasim.core.simulation.mode_choice.utilities.predictors.WalkPredictor
 import org.matsim.contribs.discrete_mode_choice.components.utils.home_finder.HomeFinder;
 import org.matsim.contribs.discrete_mode_choice.modules.config.DiscreteModeChoiceConfigGroup;
 import org.matsim.contribs.discrete_mode_choice.modules.config.VehicleTourConstraintConfigGroup;
-import org.matsim.contribs.discrete_mode_choice.replanning.time_interpreter.TimeInterpreter;
 import org.matsim.core.router.TripRouter;
+import org.matsim.core.utils.timing.TimeInterpretation;
 import org.matsim.facilities.ActivityFacilities;
 
 import com.google.inject.Provider;
@@ -87,7 +87,7 @@ public class EqasimModeChoiceModule extends AbstractEqasimExtension {
 	@Provides
 	public ModalUtilityEstimator provideModularUtilityEstimator(TripRouter tripRouter, ActivityFacilities facilities,
 			Map<String, Provider<UtilityEstimator>> factory, EqasimConfigGroup config,
-			TimeInterpreter.Factory timeInterpreterFactory, DiscreteModeChoiceConfigGroup dmcConfig) {
+			TimeInterpretation timeInterpretation, DiscreteModeChoiceConfigGroup dmcConfig) {
 		Map<String, UtilityEstimator> estimators = new HashMap<>();
 
 		for (Map.Entry<String, String> entry : config.getEstimators().entrySet()) {
@@ -101,7 +101,7 @@ public class EqasimModeChoiceModule extends AbstractEqasimExtension {
 			}
 		}
 
-		return new ModalUtilityEstimator(tripRouter, facilities, estimators, timeInterpreterFactory,
+		return new ModalUtilityEstimator(tripRouter, facilities, estimators, timeInterpretation,
 				Collections.emptySet()); // Here we may add "pt" etc. as pre-routed modes.
 	}